Bug 693507: check for out-of-bounds access
Report error in get_next_word() functions on out-of-bounds access and
handling of this error in both arithmetic and huffman decoders.

When decoding a corrupted document, the arithmetic decoder can hit the end
of the Jbig2 stream. The problem was that get_next_word() returned 0 in
that case and that this lead to an infinite loop.