Bug 694362: More tweaking in jbig2dec.
Following more discussions with Simon, and testing within MuPDF, it
appears that the original code was correct. I am therefore restoring
the original behaviour, except with a direct jump to cleanup rather
than relying on dropping into the following case.

This will result in just 1 error being given, and is cleaner, to my
mind.