jbig2dec - refine test for "Denial of Service" images
Bug 697682 "Incorrectly displayed a PDF file on GSView, MuPDF and GhostScript"

In commit 7366747076f3b75def52079bd4d5021539a16394 Simon introduced a
test to try and detect a JBIG2 compressed image which would take an
unreasonable amount of time to process. Unfortunately, the test also trips
up this real world (if rather extreme) example.

The test looks for an image to be > 16MB and for the segment being
processed to contain less than 1/256 of the image data. However, since
we use this same section for symbol decoding, the segment size can be
considerably smaller than that, and still be legal.

This commit alters the image size to 64MB (as this is, frankly, ridiculously
large for a JBIG2 image) and for the segment size to be less than
1/65536 of the total image data. This means that images less than 64MB,
with segments larger than ~1KB might still cause a DOS, but this seems
to be simply a problem we have to accept.

Since the DOS time is dependent on the size of the data, and the size
of the segment, this is a compromise.