jbig2dec: Move text region decoder parameters initialization.
Previously the text region decoder parameters were only initialized
once a text region was about to be decoded. Now they are initialized
at the beginning of the decoding of the symbol dictionary. This
prepares for sharing of the IAID, IARDX and IARDY arithmetic
integer decoder contexts as well as the SBHUFFRDX, SBHUFFRDY and
SBHUFFRSIZE huffman tables between the text region decoder and the
symbol dictionary parser itself.