Fix jbig2dec compilation when inline is not defined.
VS 2008 defines __inline but not inline, later versions of Visual Studio
do define inline, but apparently also permit __inline (presumably for
backwards compatibility).

Since we now use inline for performance purposes in jbig2_image.h its
vital that inline is defined. Here we modify the win32 config file to
define inline in terms of __inline for VS 2008 and above, unless inline
is already defined in which case we use it as is. For VS 2005
Ghostscript defines inline as nothing, so we do the same.

Finally, add a last-ditch definition of inline (as nothing) to jbig2.h
to cover any build environment where inline is still undefined, so that
the code will at least compile.