jbig2dec: Only set prev and prevprev line pointers if valid.
Previously these pointer to the previous and previous previous
line were always set, even if they underflowed. This didn't really
matter because of the added checks for the y coordinate, ensuring
that out of bounds pointers were never referenced.

Now the pointers are only assigned when valid.

This fixes Coverity issues 427566 and 427616.